Huge-ish patch.
New map, and a new potential gameplay element: Castle storming.

At some point, all the normal mobs of Clonk are easy for players.  What
is the player supposed to do, now that everything's easy?  Form a
alliance, and get a fort, and hold it against the other players!  This
will only work well if the game gets to be more of an MMO.  :P